feat: move fe calls from client-side to server-side
